add more pact handling
