+++ .envrc
@@ -3,3 +3,5 @@ for d in lambda_functions/v*/
 do
     find "${d}/functions" -type f -print0 | sort -z | xargs -0 sha1sum | sha1sum | cut -c 1-15 > "${d}"/directory_sha
 done
+
+export ENVIRONMENT="local"
