+++ lambda_functions/v1/directory_sha
@@ -1,1 +1,1 @@
-035a27a84f8f545
+ab49e17d9deadd9
