+++ lambda_functions/v1/openapi/lpa-codes-openapi-v1.yml
@@ -15,7 +15,7 @@ paths:
     get:
       summary: "Healthcheck endpoint"
       description: "Checks health of our lambda"
-      operationId: app.healthcheck
+      operationId: api.resources.handle_healthcheck
       security:
         - sigv4: []
       x-amazon-apigateway-request-validator: "all"
@@ -35,7 +35,7 @@ paths:
     post:
       summary: "Create a new code"
       description: "Create a new 12 digit alpha-numeric code unique to this LPA and Actor"
-      operationId: app.create
+      operationId: api.resources.handle_create
       security:
         - sigv4: []
       x-amazon-apigateway-request-validator: "all"
@@ -134,7 +134,7 @@ paths:
     post:
       summary: "Validate a code"
       description: "Validate a code"
-      operationId: app.validate
+      operationId: api.resources.handle_validate
       security:
         - sigv4: []
       x-amazon-apigateway-request-validator: "all"
@@ -224,7 +224,7 @@ paths:
     post:
       summary: "Revoke a code"
       description: "Revoke a code"
-      operationId: app.revoke
+      operationId: api.resources.handle_revoke
       security:
         - sigv4: []
       x-amazon-apigateway-request-validator: "all"
@@ -307,7 +307,7 @@ components:
       name: Authorization
       in: header
       x-amazon-apigateway-authtype: awsSigv4
-      x-apikeyInfoFunc: app.apikey_auth
+      x-apikeyInfoFunc: lpa_codes_mock.apikey_auth
   schemas:
     CodeCreated:
       properties:
