+++ lambda_functions/v1/requirements/pact-requirements.txt
@@ -2,4 +2,3 @@ connexion
 swagger-ui-bundle
 requests
 boto3
-moto
